switching to use TemplateID for template id's & fixing tests
